kyc-backend/requirements.txt
flask
flask-cors
pandas
openpyxl
reportlab
pyodbc
argon2-cffi
cryptography
